requirements.txt
discord.py
pynacl
youtube_dl